refactored StudentModel, CourseModel, UserModel
these classes will pass string (query) to a method in the DB class
(getData). getData() will query and execute it and write data into
List<Map<String, Objects>> and return it.

The 3 Model classes will take this List and create the proper objects
and return them to the method that originally called them.